+++ continual/utils.py
@@ -1,9 +1,9 @@
-from collections import OrderedDict
+from collections import OrderedDict, abc
 from contextlib import contextmanager
 from functools import partial, reduce
 from inspect import getsource
 from numbers import Number
-from typing import Tuple, Union
+from typing import Any, List, Sequence, Tuple, Union
 
 import torch
 from torch import Tensor, nn
@@ -216,3 +216,20 @@ def function_repr(fn):
         s = s[:i]
         s = s.rstrip()
     return s
+
+
+def flatten(lst: Union[Sequence[Any], Any], remove_none=True) -> List[Any]:
+    lst = _flatten(lst)
+    if remove_none:
+        lst = [i for i in lst if i is not None]
+    return lst
+
+
+def _flatten(lst: Union[Sequence[Any], Any]) -> List[Any]:
+    if isinstance(lst, abc.Sequence):
+        if len(lst) == 0:
+            return []
+        first, rest = lst[0], lst[1:]
+        return _flatten(first) + _flatten(rest)
+    else:
+        return [lst]