+++ CHANGELOG.md
@@ -14,6 +14,8 @@ From v1.0.0 and on, the project will adherence strictly to Semantic Versioning.
 
 ### Added
 - Add about info to package root.
+- purely functional `_forward_step` function to all modules.
+- `_state_shape` and `_dynamic_state_inds` properties to modules.
 
 ### Changed
 - Change call_mode internally from enum to tensor.